Explicit type cast to resolve different type on darwin

+++ apps/files/ajax/diskspace.go
@@ -19,7 +19,7 @@ func getFreeDiskSpace() (int64, error) {
 	if err != nil {
 		return 0, err
 	}
-	space := int64(stats.Bavail) * stats.Bsize // unused blocks * the size of one block
+	space := int64(stats.Bavail) * int64(stats.Bsize) // unused blocks * the size of one block
 	log.Debugf("available space: %v bytes", space)
 	return space, nil
 }
